fix(kb): FK author_user_id/user_id compatível com users.id legacy (#940)
users.id é int unsigned (UltimatePOS legacy), não bigint unsigned. As 6
colunas *_user_id das migrations KB declaravam unsignedBigInteger e
falhavam com SQLSTATE[HY000]:3780 "Referencing column ... incompatible".

Resultado pré-fix: módulo KB não bootstrap em DB MySQL real — 9 das 12
migrations não aplicavam.

Fix mínimo: unsignedBigInteger → unsignedInteger nas 6 referências a
users.id. business_id já estava unsignedInteger. source_doc_id (refs
mcp_memory_documents.id bigint unsigned) mantido bigint.

Validado: php artisan module:migrate KB → 12/12 DONE.

Refs: PR #934 (KB Unificado), ADR 0150, SCHEMA-DB-V1.md §3

+++ Modules/KB/Database/Migrations/2026_05_15_100005_create_kb_paths_table.php
@@ -30,7 +30,8 @@ return new class extends Migration {
             $table->unsignedSmallInteger('hue')->default(240);
             $table->string('status', 40)->default('published')
                 ->comment('draft|published|archived');
-            $table->unsignedBigInteger('author_user_id')->nullable();
+            // refs users.id (int unsigned, UltimatePOS legacy)
+            $table->unsignedInteger('author_user_id')->nullable();
             $table->timestamps();
             $table->softDeletes();
